added game over feature

+++ src/main/java/tetris/TetrisFrame.java
@@ -3,13 +3,16 @@ package tetris;
 import javax.swing.*;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
 
 public class TetrisFrame extends JFrame {
-    private final TetrisGame tetris;
-    private final TetrisView view;
+    private TetrisGame tetris;
+    private TetrisView view;
     private final TetrisKeyListener keyListener;
+    private ScheduledExecutorService scheduler;
+    private ScheduledFuture<?> scheduledFuture;
 
     public TetrisFrame(TetrisGame tetrisGame, TetrisView tetrisView, TetrisKeyListener tetrisKeyListener){
         super();
@@ -37,12 +40,38 @@ public class TetrisFrame extends JFrame {
      */
     private void scheduleDropShape() {
         int delay = 500;
-        final ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
+        scheduler = Executors.newScheduledThreadPool(1);
         final Runnable dropShapeTask = () -> {
-            tetris.drop();
-            view.repaint();
+            if(tetris.isGameOver()){
+                gameOver();
+                scheduledFuture.cancel(true);
+            } else {
+                tetris.drop();
+                view.repaint();
+            }
         };
+        scheduledFuture = scheduler.scheduleAtFixedRate(dropShapeTask, delay, delay, TimeUnit.MILLISECONDS);
+    }
 
-        scheduler.scheduleAtFixedRate(dropShapeTask, delay, delay, TimeUnit.MILLISECONDS);
+    public void gameOver(){
+        JOptionPane jOPane = new JOptionPane();
+        jOPane.createDialog("Game Over :-(");
+        int answer = jOPane.showConfirmDialog(null, "Game Over.\n Would you like to play again?",
+                "Game Over.", JOptionPane.YES_NO_OPTION);
+        if(answer == JOptionPane.YES_OPTION){
+            startNewGame();
+        }
+        else{
+            System.exit(0);
+        }
+    }
+
+    private void startNewGame() {
+        tetris = new TetrisGame();
+        view.setTetrisGameInstance(tetris);
+        keyListener.setTetrisGameInstance(tetris);
+        tetris.init();
+        view.repaint();
+        scheduleDropShape();
     }
 }
